:recycle: Change how colours are defined and used
Colours used to be defined in a Sass map, keyed by colour name, which was then manually mapped to semantically named CSS custom properties. Colours are now defined as variables, which are mapped to a semantically named Sass map.

It's a tiny change, but it means that other elements of the
CSS, like the colour overrides, work in the same way as other
semantically named colours—instead of being fixed on all
colour modes.