:memo: Update CV page
Detachs the CV page from the other site-wide styles, as it tends to get a bit left behind (and broken) with time
due to its 'not really a page on the site' schtick.

Fixes some of the out of date information.

Removes skills because no one tends to care about those so much and this is long enough as it is.